Move task_1_exec.go and task_2_exec.go into sample directory in task
