+++ client/go/startclient/startclient.go
@@ -2,12 +2,12 @@ package main
 
 import (
 	"conductor"
-	"conductor/task"
+	"conductor/task/sample"
 )
 
 func main() {
 	c := conductor.NewConductorWorker("http://localhost:8080/api", 1, 10000)
 
-	c.Start("task_1", task.Task_1_Execution_Function, false)
-	c.Start("task_2", task.Task_2_Execution_Function, true)
+	c.Start("task_1", sample.Task_1_Execution_Function, false)
+	c.Start("task_2", sample.Task_2_Execution_Function, true)
 }
